move furniture updating into FurnitureUpdater sections
