+++ src/main/resources/settings.yml
@@ -181,6 +181,8 @@ ItemUpdater:
   override_renamed_items: false
   override_lore: true # If lore should be overriden, disable if other plugins modify the OraxenItem's lore
   update_items_on_reload: true # If Oraxen should update all the items of online players on /oraxen reload items/all
+
+FurnitureUpdater:
   update_furniture_on_reload: true # If furniture should be updated on /oraxen reload items/all
   update_furniture_on_load: true # If Oraxen should update the FurnitureItem whenever an entity is loaded
   # If Oraxen should update the type if the config changes, i.e ITEM_FRAME -> DISPLAY_ENTITY
